Enable cancellation of ongoing TCP connect through a cancellation socket.
Previously, it was not possible to cancel an ongoing TCP connect until the configured timeout elapsed. With this change, it is possible to cancel an ongoing TCP connect before the timeout has elapsed.

This is implemented by providing an extra client specified socket to the DUL_ASSOCIATESERVICEPARAMETERS which are passed on to the requestAssociationTCP. The select() statement will now wake up either if a connection was established, or if data is written to the cancel socket. The client can now provide the cancel socket and send data to this socket to interrupt the connect attempt.